+++ configx/configfile_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"decipher.com/object-drive-server/configx"
+	configx "decipher.com/object-drive-server/configx"
 
 	"gopkg.in/yaml.v2"
 )
@@ -17,23 +17,58 @@ func TestMain(m *testing.M) {
 
 func TestParseWhitelistFromConfigFile(t *testing.T) {
 
-	contents := readAllOrFail("testfixtures/testconf.yml", t)
-	var conf config.ConfigFile
+	contents := readAllOrFail(t, "testfixtures/testconf.yml")
+	var conf configx.AppConfiguration
 	err := yaml.Unmarshal(contents, &conf)
 	if err != nil {
 		t.Errorf("Could not unmarshal yaml config file: %v\n", err)
 	}
 
-	if len(conf.Whitelisted) != 3 {
+	if len(conf.Whitelist) != 3 {
 		t.Fail()
 	}
-	if conf.Whitelisted[0] != "first" {
+	if conf.Whitelist[0] != "first" {
 		t.Fail()
 	}
 
 }
+func TestParseAppConfigurationFromConfigFile(t *testing.T) {
+	// os.Getenv() save out value to test cascade
+	reset1 := unsetReset("OD_DB_PORT")
+	defer reset1()
 
-func readAllOrFail(path string, t *testing.T) []byte {
+	reset2 := unsetReset("OD_EVENT_ZK_ADDRS")
+	defer reset2()
+
+	contents := readAllOrFail(t, "testfixtures/complete.yml")
+	var conf configx.AppConfiguration
+	err := yaml.Unmarshal(contents, &conf)
+	if err != nil {
+		t.Errorf("Could not unmarshal yaml config file: %v\n", err)
+	}
+
+	if conf.DatabaseConnection.Driver != "mysql" {
+		t.Errorf("expected mysql, got: %v", conf.DatabaseConnection.Driver)
+	}
+	if conf.AACSettings.CAPath != "foo" {
+		t.Errorf("expected foo, got: %v", conf.AACSettings.CAPath)
+	}
+
+	if conf.DatabaseConnection.Port != "9999" {
+		t.Errorf("expected 9999, got %v", conf.DatabaseConnection.Port)
+	}
+
+	if len(conf.EventQueue.ZKAddrs) != 2 {
+		t.Errorf("expected zk_addrs string slice of len 2, got: %v", conf.EventQueue.ZKAddrs)
+	}
+
+	if conf.Whitelist[0] != "foo" {
+		t.Errorf("expected whitelist entry foo but got: %s", conf.Whitelist[0])
+	}
+
+}
+
+func readAllOrFail(t *testing.T, path string) []byte {
 	f, err := os.Open(path)
 	if err != nil {
 		t.Fail()
@@ -53,10 +88,18 @@ func TestNormalization(t *testing.T) {
 	checklist["CN=twl-server-generic2, OU=DAE, OU=DIA, OU=twl-server-generic2, O=U.S. Government, C=US"] = "cn=twl-server-generic2,ou=dae,ou=dia,ou=twl-server-generic2,o=u.s. government,c=us"
 
 	for startingValue, expected := range checklist {
-		actual := config.GetNormalizedDistinguishedName(startingValue)
+		actual := configx.GetNormalizedDistinguishedName(startingValue)
 		if actual != expected {
 			t.Logf("Normalized %s to %s. Expected %s", startingValue, actual, expected)
 			t.Fail()
 		}
+	}
+}
+
+func unsetReset(env string) func() {
+	original := os.Getenv(env)
+	os.Setenv(env, "")
+	return func() {
+		os.Setenv(env, original)
 	}
 }